Implement Ace Value
I have made a slight deviation from the exercise. Since I believe that comparisons should be true to the cards suit and rank (i.e. a card with a lower rank should not be stated as "greater than" a card with a higher rank), I have simply changed the rank of an ace to be higher than the rank of any other card. This should keep comparisons consistent.